First Commit
Created complete code structure and base of the app. Initialized the app and created the required model for now. Also the basic navbar and footer and dummy buttons are set. Created one view with supported template and so will go on in next commit with all other necessary views and templated for complete CRUD and frontend.  Also , already created App on Heroku and declared its url in allowed host, though its useless until deployment.