Login system and authentication applied. Also user role either as manager or simple_user applied.
In this commit, login system is applied to the app. No one can access app other pages through bad practices without logging in. The app is for users of 2 separate roles either as manager or as a simple_user. The simple user can only add and view employee after logging in. Whereas , the user with manager role can perform complete crud in the app for employees. User can logout if he wants and his session will be expired. Depending on the user role, he will have either access to edit and amployee and delete or it or he will not have. If manager, then he can. If he is not the manager then he can not perform complete crud. Readme.md file is updated accordingly. Views.py has some new views for login functionality. and @login_required and @manager_required is also added for authentication purposes. Index.html file is also updated for adding the condition on depending on the user role in order to provide him access to complete crud or not. login.html template is created for the new view for login. Also the main.html file is updated for logout functionality and its view is also created in views.py. Settings.py file is also updated slightly and urls.py file is also updated for the new views and templates.